feat: remove admin-panel
