--- includes/icon.b64
@@ -1,1 +1,0 @@
-PHN2ZyB4bWxucz0iaHR0cDovL3d3dy53My5vcmcvMjAwMC9zdmciIHdpZHRoPSIyMCIgaGVpZ2h0PSIyMCIgeG1sbnM6dj0iaHR0cHM6Ly92ZWN0YS5pby9uYW5vIj48cGF0aCBmaWxsPSIjZmZmIiBkPSJNOC4zMDcuOTYzTDkuNzE1LjA3OGMuMTc5LS4xMDYuMzY4LS4xMDguNTcyLjAyM2wyLjc4MiAxLjc1NGMuMzUxLjIyMS41NTIuNTIzLjUyLjk1M2EuNTguNTggMCAwIDAgLjAwMy4xMDRjLjAwMi4wMTYuMDEyLjAzMi4wMjkuMDcybC41My0uMTM3IDQuOTc2LTEuNTQ4LjM4Ni0uMTE5Yy4xOTQtLjA1MS4zNTctLjAwNC40NDUuMTg4LjA5My4yMDIuMDIzLjUyOS0uMTM2LjU4N2wtMS4xNDkuMzg4LTIuNjEzLjc5Mi0yLjYwNi44MTljLS4xOTMuMDU4LS40MDQuMDU2LS42MjguMDg0di43MDZjMS4yNDYuMTggMi40MDkuNjY1IDMuNjA2IDEuMDMxbDMuMTE5Ljk0M2MuMjEuMDY1LjQyOC4xNC40MjkuNDMuMDAxLjMzLS4yNDEuNTIzLS41Ni40MjZsLTIuNDg5LS43Ny0zLjM0NS0xLjA0M3YuMzY5YzAgLjQ4NC0uMTkyLjc3Mi0uNjQ3Ljk1NC0uMTY3LjA2Ny0uMTg0LjE1MS0uMTQxLjMzMmwuOTk2IDQuMjkxLjU2OSAyLjQxMi41ODMgMi41NDMuNjE1IDIuNjgyYy4xMTMuNDM4LS4wMzcuNjI0LS40OTIuNjI0SDUuMTk5Yy0uMTMgMC0uMjYuMDAxLS4zOSAwLS4yODYtLjAwMy0uNDc4LS4yMzYtLjQxMy0uNTE2bC43NzUtMy4zMjguNzgxLTMuNDI5LjYyMy0yLjYxMS42NTItMi43NmMuMDEzLS4wNTUtLjAzMS0uMTctLjA3Mi0uMTgzLS42MzgtLjE5NS0uNzg1LS42ODgtLjc5Mi0xLjI3MiAwLS4wMjUtLjAwOS0uMDUtLjAxNi0uMDg5bC0yLjc1Ljg2LTMuMDE5LjkzOGMtLjIzLjA3MS0uNDk5LS4wNTctLjU1MS0uMjUxLS4wNzItLjI2OS4wMzQtLjQ3Ny4yOTgtLjU1OWw2LjI3OC0xLjk1OWMuMTczLS4wNTMuMzc3LS4wMDcuNTgtLjAwN3YtLjc1OGMtMS4wMjUtLjA5NS0xLjk1LS41MjktMi45MTYtLjgxM2wtMy45MzktMS4yYy0uMjYtLjA4LS4zODktLjM1LS4yOTktLjYyMy4wNTctLjE3Mi4zMTItLjI3LjUzNS0uMjAybDMuMjE0Ljk5MSAyLjQzOC43NThjLjAzOS4wMTIuMDgyLjAxMi4xNDkuMDIyLS4wMzEtLjgyOC41OTItMS4xNDcgMS4xNTctMS41MTJsLjc4Ny0uNTAxeiIvPjwvc3ZnPg==
