+++ inseri-core.php
@@ -28,20 +28,6 @@
  */
 
 require_once plugin_dir_path(__FILE__) . 'includes/utils.php';
-require_once plugin_dir_path(__FILE__) . 'includes/db.php';
-require_once plugin_dir_path(__FILE__) . 'includes/rest_api.php';
-require_once plugin_dir_path(__FILE__) . 'includes/admin_panel.php';
-
-global $wpdb;
-$inseri_core_db = new Inseri_Core_DB($wpdb);
-$inseri_core_rest = new Inseri_Core_Rest($inseri_core_db);
-
-register_activation_hook(__FILE__, [$inseri_core_db, 'setup_table']);
-
-/**
- * REST API
- */
-add_action('rest_api_init', [$inseri_core_rest, 'register_api_routes']);
 
 /**
  * main script
@@ -93,14 +79,6 @@ function inseri_core_block_init() {
 		register_block_type($block);
 	}
 }
-
-/**
- * Admin Panel
- */
-
-add_action('init', 'Inseri_Core_Admin::register_ui_script');
-add_action('admin_enqueue_scripts', 'Inseri_Core_Admin::load_script');
-add_action('admin_menu', 'Inseri_Core_Admin::add_menu');
 
 /**
  * Enable additional file MIMEs
