--- src/admin-panel/editor.scss
@@ -1,1 +1,0 @@
-@import '~prismjs/themes/prism.css';
