+++ src/components/index.tsx
@@ -5,6 +5,5 @@ export * from './InseriThemeProvider'
 export * from './SetupEditorEnv'
 export * from './StateProvider'
 export * from './DetailViewBody'
-export * from './AdminState'
 export * from './ContentTypeSelect'
 export * from './SelectWithAction'
